Move main to cli.py
